Add ability to filter multiple keys at the same time
